Moved things around and dropped old laravel
