+++ composer.json
@@ -11,10 +11,10 @@
     ],
     "require": {
         "php": "^7.0",
-        "illuminate/contracts": "5.1.*|5.2.*|5.3.*|5.4.*|5.5.*",
-        "illuminate/support": "5.1.*|5.2.*|5.3.*|5.4.*|5.5.*",
-        "symfony/debug": "^2.7|^3.0",
-        "symfony/http-foundation": "^2.7|^3.0",
+        "illuminate/contracts": "5.3.*|5.4.*|5.5.*",
+        "illuminate/support": "5.3.*|5.4.*|5.5.*",
+        "symfony/debug": "^3.1",
+        "symfony/http-foundation": "^3.1",
         "psr/log": "^1.0"
     },
     "require-dev": {
