+++ tests/ServiceProviderTest.php
@@ -21,7 +21,6 @@ use GrahamCampbell\Exceptions\ExceptionHandler;
 use GrahamCampbell\Exceptions\Filters\CanDisplayFilter;
 use GrahamCampbell\Exceptions\Filters\ContentTypeFilter;
 use GrahamCampbell\Exceptions\Filters\VerboseFilter;
-use GrahamCampbell\Exceptions\NewExceptionHandler;
 use GrahamCampbell\TestBenchCore\ServiceProviderTrait;
 
 /**
@@ -35,18 +34,7 @@ class ServiceProviderTest extends AbstractTestCase
 
     public function testExceptionHandlerIsInjectable()
     {
-        $this->assertIsInjectable($this->getHandlerClass());
-    }
-
-    protected function getHandlerClass()
-    {
-        $app = $this->app;
-
-        if (version_compare($app::VERSION, '5.3') < 0) {
-            return ExceptionHandler::class;
-        }
-
-        return NewExceptionHandler::class;
+        $this->assertIsInjectable(ExceptionHandler::class);
     }
 
     public function testJsonApiDisplayerIsInjectable()
